Add 'src' directory
