+++ .gitignore
@@ -11,8 +11,6 @@
 
 #cache
 .turbo
-.next
-.vercel
 
 # misc
 .DS_Store
@@ -23,7 +21,6 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 .pnpm-debug.log*
-
 
 # local env files
 .env*
