+++ src/background.ts
@@ -1,0 +1,2 @@
+export {}
+console.log("HELLO WORLD FROM BGSCRIPTS")
