+++ src/contents/plasmo.ts
@@ -1,0 +1,11 @@
+import type { PlasmoContentScript } from "plasmo"
+
+export const config: PlasmoContentScript = {
+  matches: ["https://www.plasmo.com/*"]
+}
+
+window.addEventListener("load", () => {
+  console.log("content script loaded")
+
+  document.body.style.background = "pink"
+})
