+++ src/style.css
@@ -1,0 +1,15 @@
+.new-tab {
+  background-color: #f0f0f0;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.1);
+  color: #333;
+  display: block;
+  font-size: 12px;
+  line-height: 16px;
+  margin: 0 0 10px;
+  padding: 10px;
+  position: relative;
+  text-decoration: none;
+  width: 100%;
+}
