+++ tsconfig.json
@@ -4,7 +4,7 @@
   "include": ["./**/*.ts", "./**/*.tsx"],
   "compilerOptions": {
     "paths": {
-      "~*": ["./*"]
+      "~*": ["./src/*"]
     },
     "baseUrl": "."
   }
